remove:型定義の移動
