+++ src/libs/firebase/store.ts
@@ -1,14 +1,31 @@
 import { addDoc, collection, getDocs, query } from "firebase/firestore";
 import { getDownloadURL } from "firebase/storage";
-import { CreateStoreListItem, db, StoreList, StoreListItem } from "./init";
+import { db } from "./init";
 import { uploadImage } from "./storage";
+
+export type CreateStoreListItem = {
+  name: string;
+  date: string;
+  title: string;
+  imageFile: File;
+};
+export type StoreListItem = {
+  name: string;
+  date: string;
+  title: string;
+  imageref?: string;
+};
+export type StoreList = StoreListItem[];
 
 // Listへの参照を返す
 export const getListRef = (uuid: string) =>
   collection(db, "Users", uuid, "List");
 
 // Listに要素を追加する
-export const createList = async (uuid: string, data: CreateStoreListItem) => {
+export const createList = async (
+  uuid: string,
+  data: CreateStoreListItem,
+): Promise<void> => {
   const storeData: StoreListItem = { ...data };
 
   if (data.imageFile) {
@@ -17,19 +34,18 @@ export const createList = async (uuid: string, data: CreateStoreListItem) => {
     );
   }
 
-  await addDoc(getListRef(uuid), {
+  addDoc(getListRef(uuid), {
     data,
   });
 };
 
 // List全取得
-export const getList = async (uuid: string) => {
+export const getList = async (uuid: string): Promise<StoreList> => {
   const ref = query(getListRef(uuid));
-
-  const listSnapShot = await getDocs(ref);
-
   const list: StoreList = [];
-  listSnapShot.forEach((doc) => list.push(doc.data() as StoreListItem));
+  getDocs(ref).then((snapshot) =>
+    snapshot.forEach((doc) => list.push(doc.data() as StoreListItem)),
+  );
 
   return list;
 };
